customer-360-flask-app/requirements.txt
# Install with: pip install -r requirements.txt
# app.run(host='0.0.0.0', port=5000, debug=True)

flask
pandas
scikit-learn
numpy
joblib
google-generativeai
markdown2
python-dotenv
pandas
thefuzz
python-Levenshtein